[Runtimes] Delete AbstractSpark classes and fix `Spark3Runtime` doc generation
[ML-3317](https://jira.iguazeng.com/browse/ML-3317)
[ML-4654](https://jira.iguazeng.com/browse/ML-4654)
